servo: Merge #6894 - style: Switch animation timestamps to be doubles instead of floats (from pcwalton:double-precision-timestamps); r=metajack
32-bit floats are not enough to hold timestamps since the epoch and
result in jank.

r? metajack

Source-Repo: https://github.com/servo/servo
Source-Revision: 92cbb9368471d12f0d5492abd7e04b16df549366

UltraBlame original commit: abc3282e855ee19252013da3179605c8e22b2284

+++ servo/components/script/layout_interface.rs
@@ -208,15 +208,15 @@ pub struct Animation {
     
     pub property_animation: PropertyAnimation,
     
-    pub start_time: f32,
+    pub start_time: f64,
     
-    pub end_time: f32,
+    pub end_time: f64,
 }
 
 impl Animation {
     
     #[inline]
-    pub fn duration(&self) -> f32 {
+    pub fn duration(&self) -> f64 {
         self.end_time - self.start_time
     }
 }
